Move analyzer client to its own file
